Aggregate.SeatArrangement

Documentation

data SeatArrangementLifecycle

Constructors

Initialising

Registering

Proposing

Confirming

Boarding

data SeatArrangement (sc :: Nat) (lc :: SeatArrangementLifecycle) where

Constructors

SeatArrangementInitialising :: forall (sc :: Nat). SeatArrangement sc 'Initialising
Initialise lifecycle, when a flight is scheduled, accepts reservations

SeatArrangementRegistering :: forall (sc :: Nat). Reservations sc -> SeatArrangement sc 'Registering
Register/unregister reservations

SeatArrangementProposing :: forall (sc :: Nat). Reservations sc -> SeatArrangement sc 'Proposing
Retrieving proposals

SeatArrangementConfirming :: forall (sc :: Nat). Proposals sc -> ConfirmableAssignments sc -> SeatArrangement sc 'Confirming
Assignment proposals available, accepting confirmations

SeatArrangementBoarding :: forall (sc :: Nat). Assignments sc -> SeatArrangement sc 'Boarding
All passengers are assigned a seat, commence boarding

data SomeSeatArrangement (lc :: SeatArrangementLifecycle) where

Constructors

SomeSeatArrangement :: forall (sc :: Nat) (lc :: SeatArrangementLifecycle). KnownNat sc => SeatArrangement sc lc -> SomeSeatArrangement lc

Operations related to Initialising phase

initialise :: NumberOfSeats -> SomeSeatArrangement 'Initialising

Creates an empty SeatArrangement, based on the NumberOfSeats

type RegisterProblem = ReservationsProblem

class CanRegister (sc :: Nat) (lc :: SeatArrangementLifecycle) where

Register reservation to seat arrangement, applicable in Initialising and Registering phase

Methods

register :: Reservation -> SeatArrangement sc lc -> Either RegisterProblem (SeatArrangement sc 'Registering)

Operations related to Registering phase

data UnregisterResult (sc :: Nat)

Constructors

InitialisingAfterRemoval (SeatArrangement sc 'Initialising)

RegsteringAfterRemoval (SeatArrangement sc 'Registering)

class CanUnregister (sc :: Nat) (lc :: SeatArrangementLifecycle) where

Unregisters reservation from seat arrangement, applicable in Registering phase

note that it can transitition to Initialising phase, when no reservations are present after unregister

Methods

unregister :: ReservationId -> SeatArrangement sc lc -> UnregisterResult sc

propose :: forall (sc :: Nat). KnownNat sc => SeatArrangement sc 'Registering -> SeatArrangement sc 'Proposing

Ends Registering phase, retrieving proposals

Operations related to Proposing phase

data ProposalProblem

Constructors

ProposalDoesNotCoverAllReservations

parseProposals :: forall (sc :: Nat). KnownNat sc => Proposals sc -> SeatArrangement sc 'Proposing -> Either ProposalProblem (SeatArrangement sc 'Confirming)

Parses proposals to SeatArrangement, on success the lifecycle transitions to Confirming

Operations related to Confirming phase

data ProposalsForReservationProblem

Constructors

ProposalsForReservationNotFound

ProposalsForReservationCorrupt

proposalsForReservation :: forall (ps :: Nat) (sc :: Nat). KnownNat ps => ReservationId -> SeatArrangement sc 'Confirming -> Either ProposalsForReservationProblem (Proposals ps)

Get available propoals for a specific reservation

data ConfirmProblem

Constructors

ConfirmProposalNotAvailable

ConfirmProposalAssignmentsOverlap

ConfirmProposalAssignmentProblem AssignmentsProblem

confirm :: forall (ps :: Nat) (sc :: Nat). (KnownNat ps, KnownNat sc) => ReservationId -> Proposal ps -> SeatArrangement sc 'Confirming -> Either ConfirmProblem (SeatArrangement sc 'Confirming)

Confirm seat assignments of reservation

Operations related to Boarding phase

data BoardProblem

Constructors

BoardNotAllowedWithUnassignedSeats

board :: forall (sc :: Nat). KnownNat sc => SeatArrangement sc 'Confirming -> Either BoardProblem (SeatArrangement sc 'Boarding)

Start boarding

Precondition: all seats needs to assigned