Extension.Refined

Documentation

pattern Refined :: forall {k} x p. x -> Refined p x

rerefine :: forall {k} (p :: k) x. Predicate p x => (x -> x) -> Refined p x -> Either RefineException (Refined p x)

parseSomeException :: RefineException -> Maybe SomeException

data Refined (p :: k) x

A refinement type, which wraps a value of type x.

Since: refined-0.1.0.0

class Typeable p => Predicate (p :: k) x where

A typeclass which defines a runtime interpretation of a type-level predicate p for type x.

Since: refined-0.1.0.0

Methods

validate :: Proxy p -> x -> Maybe RefineException

Check the value x according to the predicate p, producing an error RefineException if the value does not satisfy.

Note: validate is not intended to be used directly; instead, it is intended to provide the minimal means necessary for other utilities to be derived. As such, the Maybe here should be interpreted to mean the presence or absence of a RefineException, and nothing else.

data RefineException

An exception encoding the way in which a Predicate failed.

Since: refined-0.2.0.0

Constructors

RefineNotException !TypeRep
A RefineException for failures involving the Not predicate.

Since: refined-0.2.0.0

RefineAndException !TypeRep !(These RefineException RefineException)
A RefineException for failures involving the And predicate.

Since: refined-0.2.0.0

RefineOrException !TypeRep !RefineException !RefineException
A RefineException for failures involving the Or predicate.

Since: refined-0.2.0.0

RefineXorException !TypeRep !(Maybe (RefineException, RefineException))
A RefineException for failures involving the Xor predicate.

Since: refined-0.5

RefineSomeException !TypeRep !SomeException
A RefineException for failures involving all other predicates with custom exception.

Since: refined-0.5

RefineOtherException !TypeRep !Text
A RefineException for failures involving all other predicates.

Since: refined-0.2.0.0

data FromTo (mn :: Nat) (mx :: Nat)

A Predicate ensuring that the value is within an inclusive range.

>>> isRight (refine @(FromTo 0 16) @Int 13)
True

>>> isRight (refine @(FromTo 13 15) @Int 13)
True

>>> isRight (refine @(FromTo 13 15) @Int 15)
True

>>> isLeft (refine @(FromTo 13 15) @Int 12)
True

Since: refined-0.1.2

Constructors

FromTo
Since: refined-0.4.2

type (&&) = And :: k -> k1 -> Type infixr 3

The conjunction of two predicates.

Since: refined-0.2.0.0

refine :: forall {k} (p :: k) x. Predicate p x => x -> Either RefineException (Refined p x)

A smart constructor of a Refined value. Checks the input value at runtime.

Since: refined-0.1.0.0

unrefine :: forall {k} (p :: k) x. Refined p x -> x

Extracts the refined value.

Since: refined-0.1.0.0

success :: Maybe RefineException

An implementation of validate that always succeeds.

Examples

Expand

  data ContainsLetterE = ContainsLetterE

  instance Predicate ContainsLetterE Text where
    validate p t
      | any (== 'e') t = success
      | otherwise = Just $ throwRefineException (typeRep p) "Text doesn't contain letter 'e'".
  

Since: refined-0.5

throwRefineSomeException

Arguments

:: TypeRep
The TypeRep of the Predicate. This can usually be given by using typeRep.

-> SomeException
A custom exception.

-> Maybe RefineException

A handler for a RefineException.

throwRefineSomeException is useful for defining what behaviour validate should have in the event of a predicate failure with a specific exception.

Since: refined-0.5