Aggregate.SeatArrangement

Documentation

data SeatArrangementLifecycle

Constructors

Registered

Proposable

Solicitable

Announceable

Boarded

data SeatArrangement (sc :: Nat) (lc :: SeatArrangementLifecycle) where

Constructors

SeatArrangementRegistered :: forall (sc :: Nat). SeatArrangement sc 'Registered
Start of lifecycle, when a flight is scheduled, accepts reservations

SeatArrangementProposable :: forall (sc :: Nat). Reservations sc -> SeatArrangement sc 'Proposable
Accepts more reservations and removals

SeatArrangementSolicitable :: forall (sc :: Nat). Reservations sc -> SeatArrangement sc 'Solicitable
Reservations no longer accepted, collecting proposals

SeatArrangementAnnounceable :: forall (sc :: Nat). Proposals sc -> ConfirmableAssignments sc -> SeatArrangement sc 'Announceable
Assignment proposals available, accepting confirmations

SeatArrangementBoarded :: forall (sc :: Nat). Assignments sc -> SeatArrangement sc 'Boarded
End of lifecycle

data SomeSeatArrangement (lc :: SeatArrangementLifecycle) where

Constructors

SomeSeatArrangement :: forall (sc :: Nat) (lc :: SeatArrangementLifecycle). KnownNat sc => SeatArrangement sc lc -> SomeSeatArrangement lc

Operations related to Registered phase

mkSeatArrangement :: NumberOfSeats -> SomeSeatArrangement 'Registered

Creates an empty SeatArrangement, based on the NumberOfSeats

type SetReservationProblem = ReservationsProblem

class CanSetReservation (sc :: Nat) (lc :: SeatArrangementLifecycle) where

Add reservation to seat arrangement, applicable in Registered and Proposable phase

Methods

setReservation :: Reservation -> SeatArrangement sc lc -> Either SetReservationProblem (SeatArrangement sc 'Proposable)

Operations related to Proposable phase

data RemoveReservationResult (sc :: Nat)

Constructors

RegisteredAfterRemoval (SeatArrangement sc 'Registered)

ProposableAfterRemoval (SeatArrangement sc 'Proposable)

class CanRemoveReservation (sc :: Nat) (lc :: SeatArrangementLifecycle) where

Removes reservation from seat arrangement, applicable in Proposable phase

note that it can transitition to Registered phase, when no reservations are present after removal

Methods

removeReservation :: ReservationId -> SeatArrangement sc lc -> RemoveReservationResult sc

solicit :: forall (sc :: Nat). KnownNat sc => SeatArrangement sc 'Proposable -> SeatArrangement sc 'Solicitable

Ends Proposable phase, collecting proposals

Operations related to Solicitable phase

data ParseProposalProblem

Constructors

ProposalDoesNotCoverAllReservations

parseProposals :: forall (sc :: Nat). KnownNat sc => Proposals sc -> SeatArrangement sc 'Solicitable -> Either ParseProposalProblem (SeatArrangement sc 'Announceable)

Parses proposals to SeatArrangement, on success the lifecycle transitions to Announceable

Operations related to Announcable phase

data ProposalsForReservationProblem

Constructors

ProposalsForReservationNotFound

ProposalsForReservationCorrupt

proposalsForReservation :: forall (ps :: Nat) (sc :: Nat). KnownNat ps => ReservationId -> SeatArrangement sc 'Announceable -> Either ProposalsForReservationProblem (Proposals ps)

Get available propoals for a specific reservation

data ConfirmProblem

Constructors

ConfirmProposalNotAvailable

ConfirmProposalAssignmentsOverlap

ConfirmProposalAssignmentProblem AssignmentsProblem

confirm :: forall (ps :: Nat) (sc :: Nat). (KnownNat ps, KnownNat sc) => ReservationId -> Proposal ps -> SeatArrangement sc 'Announceable -> Either ConfirmProblem (SeatArrangement sc 'Announceable)

Confirm seat assignments of reservation

Operations related to Boarable phase

data BoardingProblem

Constructors

BoardingDisallowsUnassignedSeats

board :: forall (sc :: Nat). KnownNat sc => SeatArrangement sc 'Announceable -> Either BoardingProblem (SeatArrangement sc 'Boarded)

Start boarding

Precondition: all seats needs to assigned